+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.3.*
 
+* Fixed:    Money and decimal fields are now numerically sorted by the collection
 * Fixed:    Limited ranges can be correctly iterated using the cursor (instead of the repository)
 
 ### 1.3.17